# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "http://localhost:4444",
      "description": "Local Server"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "lms-backend",
    "description": "",
    "license": {
      "name": "UNLICENSED"
    }
  },
  "paths": {
...
    "/v1/course/{courseId}/chapter/{chapterId}/lesson/{lessonId}/lesson-content": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "lessonId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonContentDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "lessonId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonContentCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonContentDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "lesson content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
    "/v1/course/{courseId}/chapter/{chapterId}/lesson/{lessonId}/lesson-content/{id}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "lessonId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/LessonContentDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "lessonId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialLessonContentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonContentDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "lesson content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "lessonId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonContentDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "lesson content not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
    "/v1/course/{courseId}/chapter/{chapterId}/lesson": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonCreateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "chapter not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
    "/v1/course/{courseId}/chapter/{chapterId}/lesson/{id}": {
      "get": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/LessonDto"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialLessonDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "lesson not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [],
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "chapterId",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          },
          "400": {
            "description": "invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TypeGuardErrorany"
                }
              }
            }
          },
          "404": {
            "description": "lesson not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IErrorResponse404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TypeGuardErrorany": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "expected": {
            "type": "string"
          },
          "value": {},
          "fake_expected_typed_value_": {},
          "name": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "stack": {
            "type": "string"
          }
        },
        "required": [
          "method",
          "expected",
          "value",
          "name",
          "message"
        ]
      },
      "IErrorResponse404": {
        "type": "object",
        "properties": {
          "statusCode": {
            "const": 404
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "LessonContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lessonId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "contentType": {
            "oneOf": [
              {
                "const": "image"
              },
              {
                "const": "video"
              },
              {
                "const": "text"
              },
              {
                "const": "file"
              }
            ]
          },
          "url": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uri"
              }
            ]
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "sequence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "integer"
              }
            ]
          }
        },
        "required": [
          "id",
          "lessonId",
          "title",
          "description",
          "contentType",
          "url",
          "metadata",
          "sequence"
        ]
      },
      "LessonContentCreateDto": {
        "type": "object",
        "properties": {
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "contentType": {
            "oneOf": [
              {
                "const": "image"
              },
              {
                "const": "video"
              },
              {
                "const": "text"
              },
              {
                "const": "file"
              }
            ]
          },
          "url": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uri"
              }
            ]
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "description",
          "title",
          "contentType",
          "url",
          "metadata"
        ]
      },
      "PartialLessonContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "lessonId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "contentType": {
            "oneOf": [
              {
                "const": "image"
              },
              {
                "const": "video"
              },
              {
                "const": "text"
              },
              {
                "const": "file"
              }
            ]
          },
          "url": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string",
                "format": "uri"
              }
            ]
          },
          "metadata": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "sequence": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "integer"
              }
            ]
          }
        },
        "description": "Make all properties in T optional"
      },
      "LessonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "chapterId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "sequence": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "chapterId",
          "title",
          "description",
          "sequence"
        ]
      },
      "LessonCreateDto": {
        "type": "object",
        "properties": {
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "sequence": {
            "type": "integer"
          }
        },
        "required": [
          "description",
          "title",
          "sequence"
        ]
      },
      "PartialLessonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "chapterId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "sequence": {
            "type": "integer"
          }
        },
        "description": "Make all properties in T optional"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
...
}